fixed counters; now gettimeinusecs everywhere; need to change ro rdtsc if defined NDEBUG preliminary change to wgen score calculation to make it more accurate; disabled for now
git-svn-id: svn+ssh://combine-101/home/svn_bio/shrimp/trunk@886 450597fd-3c47-0410-84d8-f85f2794411d

+++ common/sw-full-cs.h
@@ -6,7 +6,7 @@
 #include "anchors.h"
 int	sw_full_cs_cleanup(void);
 int	sw_full_cs_setup(int, int, int, int, int, int, int, int, int, bool, int, int = 0);
-void	sw_full_cs_stats(uint64_t *, uint64_t *, uint64_t *);
+void	sw_full_cs_stats(uint64_t *, uint64_t *, double *);
 void	sw_full_cs(uint32_t *, int, int, uint32_t *, int, int, int,
 		   struct sw_full_results *, bool, bool, struct anchor *, int, int, int * = NULL);
